speech/requirements.txt
torch --index-url https://download.pytorch.org/whl/cpu
openai-whisper
sounddevice
numpy
scipy
